// frontend/src/api/models/marketSecuritySearchResult.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MarketSecuritySearchResultQuoteType } from './marketSecuritySearchResultQuoteType'

export interface MarketSecuritySearchResult {
  symbol: string
  name: string
  quoteType: MarketSecuritySearchResultQuoteType
  exchangeCode: string
  exchangeName: string
  /** @pattern ^[A-Z]{3}$ */
  currency: string
  /** @nullable */
  marketIdentifierCode: string | null
}